// modules/loan/modules/third-party-loans/enum/payment.interface.ts
export enum TPLPTypeView {
  DEFAULT = 'signatureTrackingDefault',
  TRANSACTION_HISTORY = 'transactionHistory',
  ST_OPERATIONS = 'signatureTrackingOperations',
  ST_MODIFY_TRANSACTIONS = 'signatureTrackingModify',
  ST_DETAIL_TRANSACTION = 'signatureTrackingDetail',
  ST_CONFIRM_TRANSACTION = 'signatureTrackingConfirmScreen'
}